fix($merge): 修复合并的问题

+++ myopia-business/business-core/device-core/src/main/java/com/wupol/myopia/business/core/device/domain/model/DeviceScreeningData.java
@@ -108,67 +108,67 @@ public class DeviceScreeningData implements Serializable {
      * 左眼柱镜
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal leftCyl;
+    private Double leftCyl;
 
     /**
      * 右眼柱镜
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal rightCyl;
+    private Double rightCyl;
 
     /**
      * 左眼轴位
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal leftAxsi;
+    private Double leftAxsi;
 
     /**
      * 右眼轴位
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal rightAxsi;
+    private Double rightAxsi;
 
     /**
      * 左眼瞳孔半径
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal leftPr;
+    private Double leftPr;
 
     /**
      * 右眼瞳孔半径
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal rightPr;
+    private Double rightPr;
 
     /**
      * 左眼等效球镜度
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal leftPa;
+    private Double leftPa;
 
     /**
      * 右眼等效球镜度
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal rightPa;
+    private Double rightPa;
 
     /**
      * 左眼球镜
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal leftSph;
+    private Double leftSph;
 
     /**
      * 右眼球镜
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal rightSph;
+    private Double rightSph;
 
     /**
      * 瞳距
      */
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal pd;
+    private Double pd;
 
     /**
      * 是否筛查(-1=未知,1=是,0=否)
